add forbidden logs, add forbidden chech

+++ src/main/java/controllers/CategoryController.java
@@ -1,7 +1,9 @@
 package controllers;
 
 import auth.CustomUserDetails;
+import exceptions.ForbiddenException;
 import exceptions.NotFoundException;
+import helpers.UserHelper;
 import models.CategoryModel;
 import models.LogModel;
 import org.json.simple.JSONObject;
@@ -19,14 +21,21 @@ import java.util.HashMap;
  * Контроллер категорий для модератора
  */
 @Controller
-@RequestMapping("/admin")
-public class AdminCategoryController {
+public class CategoryController {
 
     /**
      * Страница с деревом категорий
      */
     @RequestMapping(value = {"/categories" }, method = RequestMethod.GET)
-    public String index(Model model) {
+    public String index(
+            Model model,
+            Principal principal
+    ) {
+        CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
+        if (!UserHelper.isModerator(activeUser)) {
+            LogModel.addWarning(activeUser.getEmployeeId(), "Попытка доступа на страницу /categories без прав модератора");
+            throw new ForbiddenException("Доступ запрещен");
+        }
         try {
             ArrayList<CategoryModel> trees = CategoryModel.findAll();
             model.addAttribute("trees", trees);
@@ -53,6 +62,11 @@ public class AdminCategoryController {
             @RequestParam("position") int position,
             Principal principal
     ) {
+        CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
+        if (!UserHelper.isModerator(activeUser)) {
+            LogModel.addWarning(activeUser.getEmployeeId(), "Попытка добавления категории (/category/ajax-add-category) без прав модератора");
+            throw new ForbiddenException("Доступ запрещен");
+        }
         JSONObject result = new JSONObject();
         try {
             CategoryModel category = new CategoryModel(parent, position, title);
@@ -60,7 +74,6 @@ public class AdminCategoryController {
                 result.put("title", title);
                 result.put("id", category.getId());
                 result.put("error", false);
-                CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
                 LogModel.addInfo(activeUser.getEmployeeId(), "Добавлена категория "+title+", id="+category.getId());
             } else {
                 result.put("error", true);
@@ -87,6 +100,11 @@ public class AdminCategoryController {
             @RequestParam("title") String title,
             Principal principal
     ) {
+        CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
+        if (!UserHelper.isModerator(activeUser)) {
+            LogModel.addWarning(activeUser.getEmployeeId(), "Попытка переименования категории (/category/ajax-rename) без прав модератора");
+            throw new ForbiddenException("Доступ запрещен");
+        }
         JSONObject result = new JSONObject();
         try {
             CategoryModel model = CategoryModel.findById(id);
@@ -95,7 +113,6 @@ public class AdminCategoryController {
             model.update();
             result.put("title", title);
             result.put("error", false);
-            CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
             LogModel.addInfo(activeUser.getEmployeeId(), "Категория "+oldTitle+" переименована в "+title+", id="+id);
         } catch (SQLException e) {
             result.put("error", true);
@@ -119,7 +136,12 @@ public class AdminCategoryController {
             @RequestParam("newParentId") int newParentId,
             @RequestParam("position") int position,
             Principal principal
-            ) {
+    ) {
+        CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
+        if (!UserHelper.isModerator(activeUser)) {
+            LogModel.addWarning(activeUser.getEmployeeId(), "Попытка изменения сортировки категорий (/category/ajax-update-position) без прав модератора");
+            throw new ForbiddenException("Доступ запрещен");
+        }
         JSONObject result = new JSONObject();
         try {
             CategoryModel model = CategoryModel.findById(id);
@@ -128,7 +150,6 @@ public class AdminCategoryController {
                 model.update();
             }
             CategoryModel.updateSortingOfNode(newParentId, id, position);
-            CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
             LogModel.addInfo(activeUser.getEmployeeId(), "Изменен порядок подкатегорий у родителя с id="+newParentId);
             result.put("error", false);
         } catch (SQLException e) {
@@ -150,12 +171,16 @@ public class AdminCategoryController {
             @RequestParam("id") int id,
             Principal principal
     ) {
+        CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
+        if (!UserHelper.isModerator(activeUser)) {
+            LogModel.addWarning(activeUser.getEmployeeId(), "Попытка удаления категории (/category/ajax-delete) без прав модератора");
+            throw new ForbiddenException("Доступ запрещен");
+        }
         JSONObject result = new JSONObject();
         try {
             CategoryModel model = CategoryModel.findById(id);
             String title = model.getTitle();
             model.delete();
-            CustomUserDetails activeUser = (CustomUserDetails) ((Authentication) principal).getPrincipal();
             LogModel.addInfo(activeUser.getEmployeeId(), "Категория "+title+" удалена");
             result.put("error", false);
         } catch (SQLException e) {
